AlphaRadar/pyproject.toml
[project]
name = "AlphaRadar"
version = "0.1.0"
description = "Intelligent Quant Review & Monitoring Terminal"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "PyQt6>=6.6.0",
    "duckdb>=0.9.2",
    "polars>=0.20.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "akshare>=1.12.0",
    "yfinance>=0.2.33",
    "mplfinance>=0.12.10",
    "TA-Lib>=0.4.28",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "backtrader>=1.9.78.123",  # Note: verify specific version for py3 support or use forks if needed
    "schedule>=1.2.1",
    "openai>=1.3.0",
    "google-generativeai>=0.3.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[tool.black]
line-length = 88
target-version = ['py310']
